Add support for arbitrary API urls
Resolves #13: new user configuration for setting the URL and model id. However, since
the embeddings need to be deserialized from the response, and different
APIs return embeddings differently, the supported JSON structures are currently limited to that of
Ollama and OpenAI, and must be selected correctly from the settings.

Other features are also added such as debug mode logging (resolves #7) and toggling cost
estimation.